+++ packages/telescope-newsletter/lib/server/campaign.js
@@ -10,7 +10,7 @@ getCampaignPosts = function (postsCount) {
   var lastWeek = moment().subtract(7, 'days').toDate();
   var after = (typeof lastCampaign !== 'undefined') ? lastCampaign.finishedAt : lastWeek
 
-  var params = Posts.getSubParams({
+  var params = Posts.parameters.get({
     view: 'campaign',
     limit: postsCount,
     after: after