+++ packages/telescope-posts/lib/parameters.js
@@ -1,14 +1,18 @@
 /**
- * Gives an object containing the appropriate find
+ * Parameter callbacks let you add parameters to subscriptions 
+ * @namespace Posts.parameters
+ */
+Posts.parameters = {};
+
+/**
+ * Takes a set of terms, and translates them into a `parameter` object containing the appropriate find
  * and options arguments for the subscriptions's Posts.find()
  * @param {Object} terms
  */
-Posts.getSubParams = function (terms) {
+Posts.parameters.get = function (terms) {
 
   // add this to ensure all post publications pass audit-arguments-check
   check(terms, Match.Any);
-
-  var maxLimit = 200;
 
   // console.log(terms)
 
@@ -25,30 +29,43 @@ Posts.getSubParams = function (terms) {
   if (typeof Posts.views[view] !== 'undefined')
     parameters = Telescope.utils.deepExtend(true, parameters, Posts.views[view](terms));
 
-  // extend sort to sort posts by _id to break ties
-  Telescope.utils.deepExtend(true, parameters, {options: {sort: {_id: -1}}});
-
-  // if a limit was provided with the terms, add it too (note: limit=0 means "no limit")
-  if (typeof terms.limit !== 'undefined')
-    _.extend(parameters.options, {limit: parseInt(terms.limit)});
-
-  // limit to "maxLimit" posts at most when limit is undefined, equal to 0, or superior to maxLimit
-  if(!parameters.options.limit || parameters.options.limit === 0 || parameters.options.limit > maxLimit) {
-    parameters.options.limit = maxLimit;
-  }
-
-  // hide future scheduled posts unless "showFuture" is set to true or postedAt is already defined
-  if (!parameters.showFuture && !parameters.find.postedAt)
-    parameters.find.postedAt = {$lte: new Date()};
-
-  // filter by category if category _id is provided (unless categories parameter already specificed)
-  // NOTE: this is a temporary fix because views cannot currently be combined
-  if (!!terms.category && !parameters.find.categories) {
-    var categoryId = Categories.findOne({slug: terms.category})._id;
-    parameters.find.categories = {$in: [categoryId]};
-  }
+  // iterate over postsParameters callbacks
+  parameters = Telescope.callbacks.run("postsParameters", parameters, terms);
   
   // console.log(parameters);
 
   return parameters;
 };
+
+// Parameter callbacks
+
+// extend sort to sort posts by _id to break ties
+function breakTies (parameters, terms) {
+  return Telescope.utils.deepExtend(true, parameters, {options: {sort: {_id: -1}}});
+}
+Telescope.callbacks.add("postsParameters", breakTies);
+
+// limit the number of posts that can be requested at once
+function limitPosts (parameters, terms) {
+  var maxLimit = 200;
+  // if a limit was provided with the terms, add it too (note: limit=0 means "no limit")
+  if (typeof terms.limit !== 'undefined') {
+    _.extend(parameters.options, {limit: parseInt(terms.limit)});
+  }
+
+  // limit to "maxLimit" posts at most when limit is undefined, equal to 0, or superior to maxLimit
+  if(!parameters.options.limit || parameters.options.limit === 0 || parameters.options.limit > maxLimit) {
+    parameters.options.limit = maxLimit;
+  }
+  return parameters;
+}
+Telescope.callbacks.add("postsParameters", limitPosts);
+
+// hide future scheduled posts unless "showFuture" is set to true or postedAt is already defined
+function hideFuturePosts (parameters, terms) {
+  if (!parameters.showFuture && !parameters.find.postedAt) {
+    parameters.find.postedAt = {$lte: new Date()};
+  }
+  return parameters;
+}
+Telescope.callbacks.add("postsParameters", hideFuturePosts);