+++ packages/telescope-sitemap/lib/server/sitemaps.js
@@ -8,7 +8,7 @@ Meteor.startup(function() {
 
       if (!_.isFunction(postView)) return null;
 
-      var params = Posts.getSubParams(postView(terms));
+      var params = Posts.parameters.get(postView(terms));
       var post = Posts.findOne(params.find, {
         'fields': {'postedAt': 1},
         'sort': params.options.sort
@@ -43,7 +43,7 @@ Meteor.startup(function() {
     var postPages = {};
     _.each(["top", "new", "best"], function(key) {
       var siteUrl = Telescope.utils.getSiteUrl();
-      var params = Posts.getSubParams(Posts.views[key]());
+      var params = Posts.parameters.get(Posts.views[key]());
       var posts = Posts.find(params.find, {
         fields: {postedAt: 1, slug: 1, _id: 1},
         limit: 100,