+++ packages/telescope-tags/lib/views.js
@@ -1,4 +1,3 @@
-
 // category post list parameters
 Posts.views.add("category", function (terms) {
   var category = Categories.findOne({slug: terms.category});
@@ -9,3 +8,14 @@ Posts.views.add("category", function (terms) {
     options: {sort: {sticky: -1, score: -1}} // for now categories views default to the "top" view
   };
 });
+
+// add category parameter to publications/subscriptions
+function addCategoryParameter (parameters, terms) {
+  // filter by category if category _id is provided (unless categories parameter already specificed)
+  if (!!terms.category && !parameters.find.categories) {
+    var categoryId = Categories.findOne({slug: terms.category})._id;
+    parameters.find.categories = {$in: [categoryId]};
+  }
+  return parameters;
+}
+Telescope.callbacks.add("postsParameters", addCategoryParameter);